Static files, basic version of handbuilder
The instruction text needs some rewriting, also needs some work to be more nice to use. Maybe some feedback for whether hands are completed?

Can I load the score in a frame or load it dynamically into a div instead of redirecting to the score page?

Need fix the image generator to display open calls correctly